langchain_app/requirements.txt
blobfile
langchain
langchain-community
langchain-core
langchain-huggingface
openai
python-dotenv
sentencepiece
streamlit
transformers
tiktoken
